requirements.txt
langchain
langchain-openai
langchain-pinecone
pinecone-client
python-dotenv
tiktoken
